requirements.txt
streamlit>=1.20.0
python-pptx>=0.6.21
lxml>=4.9.0
